Switch mobile strategy from Capacitor to React Native (Expo)
Update docs/mobile-app-plan.md to v10.0: drop the Vite + TanStack Start
migration and Capacitor WebView approach in favor of a React Native (Expo)
app. Keep Next.js for web unchanged.

Extract shared business logic into cross-platform packages:
- @boardsesh/ble-protocol: Aurora v2/v3 + MoonBoard BLE protocol encoding
- @boardsesh/queue: Queue state machine, playlist suggestions, event utils

Web app imports updated to re-export from shared packages, preserving all
existing APIs.
