+++ package.json
@@ -2,7 +2,8 @@
   "name": "boardsesh-monorepo",
   "private": true,
   "workspaces": [
-    "packages/*"
+    "packages/*",
+    "packages/shared/*"
   ],
   "scripts": {
     "lint": "vp lint",
