+++ packages/shared/queue/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "@boardsesh/queue",
+  "version": "1.0.0",
+  "type": "module",
+  "main": "src/index.ts",
+  "types": "src/index.ts",
+  "exports": {
+    ".": {
+      "types": "./src/index.ts",
+      "import": "./src/index.ts",
+      "default": "./src/index.ts"
+    }
+  },
+  "dependencies": {
+    "@boardsesh/shared-schema": "*"
+  },
+  "devDependencies": {
+    "typescript": "^5.9.3"
+  }
+}
