+++ packages/shared/queue/src/index.ts
@@ -1,0 +1,33 @@
+// Pure queue state machine — no React, no DOM, works in any JS runtime.
+// React-specific wrappers (useQueueReducer) and web-specific types
+// (QueueContextType, QueueDataType, QueueActionsType) stay in the web app.
+
+export { queueReducer, initialState } from './reducer';
+
+export type {
+  QueueState,
+  QueueAction,
+  QueueSearchParams,
+  ClimbQueue,
+  PlaylistSuggestionSource,
+  SetCurrentClimbOptions,
+  AddToQueueSource,
+  PeerId,
+  UserName,
+} from './types';
+
+// Re-export shared-schema types used by consumers
+export type { Climb, ClimbQueueItem, QueueItemUser } from './types';
+
+export { insertQueueItemIdempotent, evaluateQueueEventSequence } from './event-utils';
+export type { QueueSequenceDecision } from './event-utils';
+
+export {
+  mergeUniquePlaylistClimbs,
+  playlistSuggestionSourceMatches,
+  getPlaylistSuggestedClimbs,
+  pruneSuggestedQueueItemsAfterCurrent,
+  insertQueueItemAfterCurrent,
+  getPlaylistPeekQueueItemUuid,
+  isPlaylistPeekQueueItemUuid,
+} from './playlist-suggestions';
