+++ packages/web/app/components/board-bluetooth-control/bluetooth-shared.ts
@@ -1,21 +1,21 @@
-// Shared BLE transport constants/helpers used by all board protocols.
+// Re-export pure transport constants/helpers from the shared BLE protocol package.
+export {
+  MAX_BLUETOOTH_MESSAGE_SIZE,
+  MESSAGE_BODY_MAX_LENGTH,
+  AURORA_ADVERTISED_SERVICE_UUID,
+  UART_SERVICE_UUID,
+  UART_WRITE_CHARACTERISTIC_UUID,
+  splitMessages,
+} from '@boardsesh/ble-protocol/transport';
 
-export const MAX_BLUETOOTH_MESSAGE_SIZE = 20;
-export const MESSAGE_BODY_MAX_LENGTH = 255;
-
-export const AURORA_ADVERTISED_SERVICE_UUID = '4488b571-7806-4df6-bcff-a2897e4953ff';
-export const UART_SERVICE_UUID = '6e400001-b5a3-f393-e0a9-e50e24dcca9e';
-export const UART_WRITE_CHARACTERISTIC_UUID = '6e400002-b5a3-f393-e0a9-e50e24dcca9e';
-
-export const splitMessages = (buffer: Uint8Array) =>
-  Array.from({ length: Math.ceil(buffer.length / MAX_BLUETOOTH_MESSAGE_SIZE) }, (_, i) =>
-    buffer.slice(i * MAX_BLUETOOTH_MESSAGE_SIZE, (i + 1) * MAX_BLUETOOTH_MESSAGE_SIZE),
-  );
+import { UART_SERVICE_UUID, UART_WRITE_CHARACTERISTIC_UUID } from '@boardsesh/ble-protocol/transport';
 
 // Small delay between write-without-response chunks to avoid overwhelming the BLE stack.
 // Matches the pacing strategy used in the Capacitor adapter.
 const INTER_CHUNK_DELAY_MS = 5;
 const delay = (ms: number) => new Promise<void>((resolve) => setTimeout(resolve, ms));
+
+// --- Web-specific BLE helpers (use Web Bluetooth DOM types) ---
 
 export const writeCharacteristicSeries = async (
   characteristic: BluetoothRemoteGATTCharacteristic,
