+++ packages/web/package.json
@@ -28,10 +28,12 @@
     "@atlaskit/pragmatic-drag-and-drop-react-drop-indicator": "^3.2.12",
     "@auth/drizzle-adapter": "^1.11.1",
     "@aws-sdk/client-s3": "^3.1000.0",
+    "@boardsesh/ble-protocol": "*",
     "@boardsesh/board-constants": "*",
     "@boardsesh/board-renderer-wasm": "*",
     "@boardsesh/crypto": "*",
     "@boardsesh/db": "*",
+    "@boardsesh/queue": "*",
     "@boardsesh/shared-schema": "*",
     "@emoji-mart/data": "^1.2.1",
     "@emoji-mart/react": "^1.1.1",
